# lib/api-zod/src/generated/types/soilAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CropRecommendation } from "./cropRecommendation";
import type { FertilizerRecommendation } from "./fertilizerRecommendation";
import type { SoilCorrection } from "./soilCorrection";

export interface SoilAnalysisResult {
  id: number;
  crops: CropRecommendation[];
  fertilizers: FertilizerRecommendation[];
  soilCorrections: SoilCorrection[];
  explanationEnglish: string;
  explanationTamil: string;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  languageName?: string | null;
  /** @nullable */
  temperature?: number | null;
  /** @nullable */
  humidity?: number | null;
  createdAt: string;
}