# lib/api-zod/src/generated/types/soilInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SoilInputSeason } from "./soilInputSeason";
import type { SoilInputWaterAvailability } from "./soilInputWaterAvailability";

export interface SoilInput {
  /** @nullable */
  farmerName?: string | null;
  location: string;
  season: SoilInputSeason;
  waterAvailability: SoilInputWaterAvailability;
  ph: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  moisture: number;
  /** @nullable */
  preferredCrop?: string | null;
  /**
   * BCP 47 language code, e.g. 'ta', 'hi', 'te'. Defaults to 'ta' (Tamil).
   * @nullable
   */
  language?: string | null;
  /**
   * Current ambient temperature in Celsius from GPS weather data
   * @nullable
   */
  temperature?: number | null;
  /**
   * Current relative humidity percentage from GPS weather data
   * @nullable
   */
  humidity?: number | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
}